shades doesnt like the older hsqldb because it doesnt have a LIMIT keyword, so update hibernate instead
git-svn-id: file:///home/igor/dev/stuff/svnbackup/branches/WICKET_1_2@1056 ef7698a4-5110-0410-9fc6-c7eb3693863f

+++ wicket-phonebook/src/java/wicket/contrib/phonebook/HibernateContactDao.java
@@ -106,7 +106,7 @@ public class HibernateContactDao implements ContactDao {
 	 * @return count
 	 */
 	public int count(Contact filter) {
-		return ((Integer) buildFindQuery(null, filter, true).uniqueResult())
+		return ((Long) buildFindQuery(null, filter, true).uniqueResult())
 				.intValue();
 	}
